+++ R/nod.R
@@ -14,7 +14,7 @@
 #'               nod(path="by_type")
 #'  session
 #' }
-#' @importFrom urltools domain path url_parse url_compose
+#' @importFrom httr parse_url modify_url
 #' @export
 nod <- function(bow, path, verbose=FALSE){
 
@@ -23,13 +23,14 @@ nod <- function(bow, path, verbose=FALSE){
 
   # if user supplied URL instead of path
   if(grepl("://|www\\.", path)){
-    if(urltools::domain(path)!=bow$domain)
+    url_parsed <- httr::parse_url(path)
+    url_subdomain <- paste0(url_parsed$scheme, "://", url_parsed$hostname)
+    if(url_subdomain!=bow$domain)
       bow <- bow(url = path, user_agent = bow$user_agent, delay=bow$delay, bow$config)
-    path <- urltools::path(path)
+    path <- url_parsed$path
   }
-  parsed_url <- urltools::url_parse(bow$url)
-  parsed_url$path <- path
-  bow$url <- urltools::url_compose(parsed_url)
+
+  bow$url <- httr::modify_url(bow$url, path=path)
 
   if(verbose && !is_scrapable(bow))
     warning("Psst!...It's not a good idea to scrape here!", call. = FALSE)